requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
drf-spectacular>=0.24.1,<0.25
